Add sm as a dependency of attestor-package.
Because the verifier needs a copy of the fw_payload.bin to calculate
the hash of the security monitor.

Unfortunately this adds a backward dependency from
sdk/examples/attestation to sm/, but it seems that the dependency on
the eyrie runtime is also introduced this way, so maybe it is not a
big problem for now.

In the future, we should introduce better modularization and avoid
such dependencies.